+++ docs/getting_started/metrics_example.py
@@ -75,4 +75,10 @@ histogram = meter.create_histogram("histogram")
 histogram.record(99.9)
 
 # Async Gauge
-gauge = meter.create_observable_gauge("gauge", [observable_gauge_func])
+observable_gauge = meter.create_observable_gauge(
+    "observable_gauge", [observable_gauge_func]
+)
+
+# Sync Gauge
+gauge = meter.create_gauge("gauge")
+gauge.set(1)