+++ exporter/opentelemetry-exporter-prometheus/src/opentelemetry/exporter/prometheus/__init__.py
@@ -378,6 +378,11 @@ class _CustomCollector:
         self, name: str, description: str, attributes: Dict[str, str]
     ) -> InfoMetricFamily:
         """Create an Info Metric Family with list of attributes"""
+        # sanitize the attribute names according to Prometheus rule
+        attributes = {
+            self._sanitize(key): self._check_value(value)
+            for key, value in attributes.items()
+        }
         info = InfoMetricFamily(name, description, labels=attributes)
         info.add_metric(labels=list(attributes.keys()), value=attributes)
         return info