+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/__init__.py
@@ -15,8 +15,9 @@
 
 from opentelemetry.sdk.metrics._internal import Meter, MeterProvider
 from opentelemetry.sdk.metrics._internal.exceptions import MetricsTimeoutError
+from opentelemetry.sdk.metrics._internal.instrument import Counter
+from opentelemetry.sdk.metrics._internal.instrument import Gauge as _Gauge
 from opentelemetry.sdk.metrics._internal.instrument import (
-    Counter,
     Histogram,
     ObservableCounter,
     ObservableGauge,
@@ -30,6 +31,7 @@ __all__ = [
     "MetricsTimeoutError",
     "Counter",
     "Histogram",
+    "_Gauge",
     "ObservableCounter",
     "ObservableGauge",
     "ObservableUpDownCounter",